requirements.txt
asgiref==3.2.10
deap==1.3.1
Django==3.1.1
django-cors-headers==3.5.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
numpy==1.19.2
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
